Rename module to tfa. Clean up comments.
