+++ satsim/image/augment.py
@@ -8,7 +8,7 @@ import tensorflow as tf
 from satsim.geometry.sprite import load_sprite_from_file
 from satsim.math import fftconv2p
 from satsim.image.fpa import add_counts
-import satsim.mytfa as tfa
+import satsim.tfa as tfa
 
 
 def scatter_shift_random(image, t, loc, scale, length, spacial_osf=1, normalize=True, mode='fft', interpolation='nearest', dtype=tf.float32):
