--- satsim/mytfa/image/__init__.py
@@ -1,42 +1,0 @@
-# Copyright 2019 The TensorFlow Authors. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-# ==============================================================================
-"""Additional image manipulation ops."""
-
-#from tensorflow_addons.image.distort_image_ops import adjust_hsv_in_yiq
-#from tensorflow_addons.image.compose_ops import blend
-#from tensorflow_addons.image.color_ops import equalize
-#from tensorflow_addons.image.color_ops import sharpness
-#from tensorflow_addons.image.connected_components import connected_components
-#from tensorflow_addons.image.cutout_ops import cutout
-#from tensorflow_addons.image.dense_image_warp import dense_image_warp
-#from tensorflow_addons.image.distance_transform import euclidean_dist_transform
-#from tensorflow_addons.image.dense_image_warp import interpolate_bilinear
-#from tensorflow_addons.image.interpolate_spline import interpolate_spline
-#from tensorflow_addons.image.filters import gaussian_filter2d
-#from tensorflow_addons.image.filters import mean_filter2d
-#from tensorflow_addons.image.filters import median_filter2d
-#from tensorflow_addons.image.cutout_ops import random_cutout
-#from tensorflow_addons.image.distort_image_ops import random_hsv_in_yiq
-#from tensorflow_addons.image.resampler_ops import resampler
-from satsim.mytfa.image.transform_ops import rotate
-from satsim.mytfa.image.transform_ops import shear_x
-from satsim.mytfa.image.transform_ops import shear_y
-#from tensorflow_addons.image.sparse_image_warp import sparse_image_warp
-from satsim.mytfa.image.transform_ops import compose_transforms
-from satsim.mytfa.image.transform_ops import angles_to_projective_transforms
-from satsim.mytfa.image.transform_ops import transform
-from satsim.mytfa.image.translate_ops import translate
-from satsim.mytfa.image.translate_ops import translate_xy
-from satsim.mytfa.image.translate_ops import translations_to_projective_transforms
