+++ satsim/tfa/image/transform_ops.py
@@ -1,23 +1,9 @@
-# Copyright 2019 The TensorFlow Authors. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-# ==============================================================================
 """Image transform ops."""
 
 import tensorflow as tf
-from satsim.mytfa.image import utils as img_utils
-from satsim.mytfa.utils.types import TensorLike
-from satsim.mytfa.image.utils import wrap, unwrap
+from satsim.tfa.image import utils as img_utils
+from satsim.tfa.utils.types import TensorLike
+from satsim.tfa.image.utils import wrap, unwrap
 
 from typing import Optional
